[NumberField] Refactor to useRenderElement
