+++ packages/react/src/number-field/decrement/NumberFieldDecrement.tsx
@@ -1,11 +1,11 @@
 'use client';
 import * as React from 'react';
 import { useNumberFieldRootContext } from '../root/NumberFieldRootContext';
-import { useComponentRenderer } from '../../utils/useComponentRenderer';
 import type { NumberFieldRoot } from '../root/NumberFieldRoot';
 import { useNumberFieldButton } from '../root/useNumberFieldButton';
-import { BaseUIComponentProps, HTMLProps } from '../../utils/types';
+import type { BaseUIComponentProps } from '../../utils/types';
 import { styleHookMapping } from '../utils/styleHooks';
+import { useRenderElement } from '../../utils/useRenderElement';
 
 /**
  * A stepper button that decreases the field value when clicked.
@@ -14,10 +14,10 @@ import { styleHookMapping } from '../utils/styleHooks';
  * Documentation: [Base UI Number Field](https://base-ui.com/react/components/number-field)
  */
 export const NumberFieldDecrement = React.forwardRef(function NumberFieldDecrement(
-  props: NumberFieldDecrement.Props,
+  componentProps: NumberFieldDecrement.Props,
   forwardedRef: React.ForwardedRef<HTMLButtonElement>,
 ) {
-  const { render, className, disabled: disabledProp = false, ...otherProps } = props;
+  const { render, className, disabled: disabledProp = false, ...elementProps } = componentProps;
 
   const {
     allowInputSyncRef,
@@ -66,25 +66,18 @@ export const NumberFieldDecrement = React.forwardRef(function NumberFieldDecreme
     locale,
   });
 
-  const propGetter = React.useCallback(
-    (externalProps: HTMLProps) => getCommonButtonProps(false, externalProps),
-    [getCommonButtonProps],
-  );
-
-  const { renderElement } = useComponentRenderer({
-    propGetter,
+  const element = useRenderElement('button', componentProps, {
     ref: forwardedRef,
-    render: render ?? 'button',
     state,
-    className,
-    extraProps: otherProps,
+    props: [getCommonButtonProps(false), elementProps],
     customStyleHookMapping: styleHookMapping,
   });
 
-  return renderElement();
+  return element;
 });
 
 export namespace NumberFieldDecrement {
   export interface State extends NumberFieldRoot.State {}
+
   export interface Props extends BaseUIComponentProps<'button', State> {}
 }
