+++ packages/react/src/number-field/group/NumberFieldGroup.tsx
@@ -2,10 +2,9 @@
 import * as React from 'react';
 import { useNumberFieldRootContext } from '../root/NumberFieldRootContext';
 import type { NumberFieldRoot } from '../root/NumberFieldRoot';
-import { mergeProps } from '../../merge-props';
 import type { BaseUIComponentProps } from '../../utils/types';
-import { useComponentRenderer } from '../../utils/useComponentRenderer';
 import { styleHookMapping } from '../utils/styleHooks';
+import { useRenderElement } from '../../utils/useRenderElement';
 
 /**
  * Groups the input with the increment and decrement buttons.
@@ -14,38 +13,25 @@ import { styleHookMapping } from '../utils/styleHooks';
  * Documentation: [Base UI Number Field](https://base-ui.com/react/components/number-field)
  */
 export const NumberFieldGroup = React.forwardRef(function NumberFieldGroup(
-  props: NumberFieldGroup.Props,
+  componentProps: NumberFieldGroup.Props,
   forwardedRef: React.ForwardedRef<HTMLDivElement>,
 ) {
-  const { render, className, ...otherProps } = props;
+  const { render, className, ...elementProps } = componentProps;
 
   const { state } = useNumberFieldRootContext();
 
-  const getGroupProps = React.useCallback(
-    (externalProps = {}) =>
-      mergeProps(
-        {
-          role: 'group',
-        },
-        externalProps,
-      ),
-    [],
-  );
-
-  const { renderElement } = useComponentRenderer({
-    propGetter: getGroupProps,
+  const element = useRenderElement('div', componentProps, {
     ref: forwardedRef,
-    render: render ?? 'div',
     state,
-    className,
-    extraProps: otherProps,
+    props: [{ role: 'group' }, elementProps],
     customStyleHookMapping: styleHookMapping,
   });
 
-  return renderElement();
+  return element;
 });
 
 export namespace NumberFieldGroup {
   export interface State extends NumberFieldRoot.State {}
+
   export interface Props extends BaseUIComponentProps<'div', State> {}
 }
