+++ packages/react/src/number-field/scrub-area-cursor/NumberFieldScrubAreaCursor.test.tsx
@@ -14,7 +14,6 @@ const defaultScrubAreaContext: NumberFieldScrubAreaContext = {
   direction: 'horizontal',
   pixelSensitivity: 2,
   teleportDistance: undefined,
-  getScrubAreaProps: (p = {}) => p,
   scrubAreaCursorRef: React.createRef<HTMLSpanElement>(),
   scrubAreaRef: React.createRef<HTMLDivElement>(),
 };
